Marks skipped tests as skipped, not passed.
Decorates tests that are potentially skipped, as well as broken tests
that are always skipped, as being skipped, rather than indicating
falsely that these tests have passed (the result of premature return
statements prior to any assertions in the tests).

This introduces another dependency for Python 2.6, the unittest2 module,
which back-ported this functionality from Python 2.7 and Python 3.