Configurable encoding reading compressed VCF files
The encoding used to read compressed VCF files was fixed to
ASCII, but is now configurable with the optional `encoding`
parameter to `vcf.Reader()`.

This is really a stopgap solution to read compressed VCF
files containing content that cannot be ASCII-decoded. More
changes are need to properly handle encoding/decoding issues
on both Python 2 and 3.

Fixes #201